Initial support for generating AMD javascript modules from "python" modules.
bin/rapydscript
---------------

Added the following new flags:
   --amd: Use Asynchronous Module Definition to import modules.
   --py: Specify the source file extension if different from .pyj.
         e.g. --py=.py

Read the environment variable PYTHONPATH for a list of search paths to
find
modules.

lib/ast.pyj
-----------

Capture the filname path of the module being imported.  This will be
used when
outputing as an AMD module.  The path captured will be the first file
found
in the PYTHONPATH.

lib/parse.pyj
-------------
- Refactored some common code.
- Capture the path of the module being imported in the AST_Import node.

lib/output.pyj
--------------
- Added code to output an AMD compliant javascript module.