Use isArray instead of instanceof in the compiler
See #70 for motivation
The isinstance() function will have to be special cased
in parse.pyj and output.pyj to handle this correctly.
It needs to be refactored anyway to accept a variable number of args or
a list as the single arg.

Also improve the baselib len() implementation, since I was there anyway.

All tests still pass.

...