Facility for creating polyfills in baselib
Now you can define items in baselib that are functions and the compiler
will output javascript that calls the function and assigns its result
to a variable named for the function. This can be used to create
polyfills for ES 6 features that we would like to use.

The first such polyfill is for Symbol.for() which will eventually be
used by the code dealing with python function calling conventions
(*args, **kwargs).