Implement incremental compilation
When importing modules, the compiler stores the compiled module in a
filename.pyj-cached file. The cache contains the SHA1 signature for both
the .pyj file being imported and the compiler source code. If either of
these do not match, the cache is invalidated.

The cache means that when working on a large project with many modules,
the cycle of edit one module - compile - test is greatly shortened,
since now parsing and output generation only happens for the changed
module and the main module instead of all imported modules.